requirements.txt
packaging==25.0
pip==26.0.1
pyodbc==5.3.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
python-dotenv==1.2.1
setuptools==80.10.2
wheel==0.46.3
